requirements.txt
streamlit
langchain-ollama
pdfminer.six
docx2python
beautifulsoup4
lxml
google-generativeai